requirements.txt
Flask
numpy
onnxruntime
Pillow
